+++ seqan/extras/demos/tutorial/data_journaling/solution_online_search_assignment4.cpp
@@ -86,7 +86,7 @@ void _findInOriginalNode(String<int> & hitTarget,
     if(!empty(refHits))
     {
         // Find upper bound to physical position in sorted refHits.
-        THitIterator itHit = std::upper_bound(begin(refHits),end(refHits),entriesIt->physicalPosition);
+        THitIterator itHit = std::upper_bound(begin(refHits),end(refHits),(int)entriesIt->physicalPosition);
         // Make sure we do not miss hits that begin at physical position of current node.
         if(itHit != begin(refHits) && *(itHit - 1) >= (int)entriesIt->physicalPosition)
             --itHit;
